requirements.txt
pandas==1.3.3
requests==2.26.0
psycopg2-binary==2.9.1
sqlalchemy==1.4.25
apache-airflow==2.2.3